Centralize specific code into apps
