+++ ahmia/search/urls.py
@@ -3,12 +3,12 @@ from django.conf.urls import patterns
 
 urlpatterns = patterns(
     '',
-    (r'^search/', 'search.views.searchengine.results'), # results
-    (r'^i2p_search/', 'search.views.searchengine.results'), # results
+    (r'^search/', 'search.views.results'), # results
+    (r'^i2p_search/', 'search.views.results'), # results
 )
 
 # Elasticsearch API
 urlpatterns += patterns(
     '',
-    (r'^elasticsearch/', 'search.views.searchengine.proxy'),
+    (r'^elasticsearch/', 'search.views.proxy'),
 )
