+++ ahmia/stats/urls.py
@@ -4,11 +4,11 @@ from django.conf.urls import patterns
 urlpatterns = patterns(
     '',
     # Stats
-    (r'^stats/viewer', 'search.views.stats.statsviewer'),
-    (r'^stats/popularity', 'search.views.stats.stats'),
-    (r'^stats/tor2web', 'search.views.stats.tor2web'),
-    (r'^stats/history', 'search.views.stats.history'),
-    (r'^stats/traffic', 'search.views.stats.trafficviewer'),
-    (r'^stats/services', 'search.views.stats.services'),
-    (r'^stats/onionsovertime', 'search.views.stats.onionsovertime'),
+    (r'^stats/viewer', 'stats.views.statsviewer'),
+    (r'^stats/popularity', 'stats.views.stats'),
+    (r'^stats/tor2web', 'stats.views.tor2web'),
+    (r'^stats/history', 'stats.views.history'),
+    (r'^stats/traffic', 'stats.views.trafficviewer'),
+    (r'^stats/services', 'stats.views.services'),
+    (r'^stats/onionsovertime', 'stats.views.onionsovertime'),
 )
